Refactor product hero gallery and breadcrumbs
